Separate api file gspn_transformation
